Format quick_win score as integer across pipeline and UI
- Ensure quick_win formula computes integer values in eval_api
- Update IssueView model and repository mappings to coerce quick_win to int
- Format score_badge component to render whole integers for display, titles, and aria labels
- Update test cases and assertions

+++ craft_dashboard/models/views.py
@@ -30,7 +30,7 @@ class IssueView:
     actionability: float | None = None
     complexity: float | None = None
     impact: float | None = None
-    quick_win: float | None = None
+    quick_win: int | None = None
     confidence: float | None = None
     has_related_links: bool = False
